refactor: shared 패키지 위치 변경 (infrastructure 내부로 이동)
