+++ src/main/java/sns/pinocchio/infrastructure/shared/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package sns.pinocchio.shared.exception;
+package sns.pinocchio.infrastructure.shared.exception;
 
 import io.swagger.v3.oas.annotations.Hidden;
 import java.util.List;
@@ -7,7 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import sns.pinocchio.shared.response.GlobalApiResponse;
+import sns.pinocchio.infrastructure.shared.response.GlobalApiResponse;
 
 @Hidden
 @RestControllerAdvice(
