+++ src/main/java/sns/pinocchio/infrastructure/shared/response/GlobalApiResponse.java
@@ -1,4 +1,4 @@
-package sns.pinocchio.shared.response;
+package sns.pinocchio.infrastructure.shared.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
